Make lyric row editor into individual class.
