+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/EditLyricRow.cs
@@ -3,13 +3,10 @@
 
 #nullable disable
 
-using osu.Framework.Allocation;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Info;
 using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Screens.Edit;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
 {
@@ -40,41 +37,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
                 Margin = new MarginPadding { Left = 10 },
                 RelativeSizeAxes = Axes.X,
             };
-        }
-
-        public class SingleLyricEditor : Container
-        {
-            [Cached]
-            private readonly EditorKaraokeSpriteText karaokeSpriteText;
-
-            public SingleLyricEditor(Lyric lyric)
-            {
-                CornerRadius = 5;
-                AutoSizeAxes = Axes.Y;
-                Padding = new MarginPadding { Bottom = 10 };
-                Children = new Drawable[]
-                {
-                    karaokeSpriteText = new EditorKaraokeSpriteText(lyric),
-                    new TimeTagLayer(lyric)
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    new CaretLayer(lyric)
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    new BlueprintLayer(lyric)
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    }
-                };
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(EditorClock clock)
-            {
-                karaokeSpriteText.Clock = clock;
-            }
         }
     }
 }
